adding package setup
